More detail about Project Director requirements
There are a number of legal and other eligibility requirements for
Project Directors. These have caused significant confusion during passed
elections, so we want to at least document the status quo.

The edits here attempt to clarify the requirements and how we've
interpreted them in practice, as well as provide more background on the
fact that some of them are rooted in the Foundation's legal
responsibilities.

Issue #66