Allow messages from observers during meetings
The original observer policy is written in such a way that any message
sent by an observer, except during check-in and check-out, is a
violation of the policy.

However, this policy is often violated in practice, and such
violations often add useful context that is helpful to the council.
The fact that these are violations puts the facilitator in the awkward
position of either ignoring a piece of useful information that the
council should consider or seeming to validate a violation of the
policy by acknowledging it.

These violators are -- in all cases to date -- respected project
members who are adding this context in good faith and with the
intention of helping the council.  Consequently, we've not had much
apparent appetite for enforcing the policy by reprimanding such
members or banning them from future meetings.

This is all kind of a silly situation to be in, so let's change the
policy to explicitly allow for such messages to be sent, while simply
asking observers to be judicious in sending these messages so as to
avoid distractions.