Add "of immediate concern" to context encouragement
If we're allowing observers to send messages, we should suggest when
such messages should be sent rather than just saying when they should
not, so as to give some sense of balance.  E.g., if we're about to
make a decision based on information the observer knows to be
incorrect, that would be a good time to send a message.  Or if we seem
blocked on missing some information we need, and the observer knows
the answer, that would be a good time.

So let's say, "important context that may be helpful to the council
*and is of immediate concern*" to try to be as precise as possible
about this.