pd: elaborate on responsibilities
The responsibilities of the Project Directors are unclear and leave a
lot of ambiguity:

- Who is ultimately responsible for the Project's relationship with the
  Foundation?
- What decisions can the PDs make independently of consultation with the
  LC?
- Who should the Foundation contact to understand the position of the Rust
  Project?
- What are the LC's expectations of the PDs?
- Who should the Foundation listen to if the PDs disagree with the LC?
  - The LC are the governing body of the Rust project so ultimately
    determine the project's view, but as board members, the PDs are the
    highest leadership body of the Foundation, so who takes priority?
- Do the PDs have any responsibilities within the Rust project?

The Foundation needs clear and coherent communication from the Rust
project in order to effectively respond to the Project's needs. Ambiguity
in the responsibility for the Project's communication with the Foundation,
and between the PDs and LC, increases the risk of miscommunication,
misunderstanding, and distrust.

Given the broad remit of the LC, the varied operations of the Rust
Foundation and many responsibilities of the PDs: it is not practical
for the LC to be directly involved in all of the decisions which must
be made by PDs, or for PDs to consult the LC on every decision made.

Some decisions are simply too in-the-weeds or unrelated to the Project
to require involvement of the LC, for example: PDs could be asked to
provide feedback to the Foundation on their proposed requirements for
associate membership - an important issue but one where PDs could be
entrusted to act independently in line with their best judgement and
understanding of the LC's priorities and views.

Feedback or decisions may be requested of PDs during board meetings
where it simply isn't possible for the PDs to ask the LC and where they
must be entrusted to act independently.

However, there are larger items that will unfold over the course of
several board meetings, such as how best to support maintainers, where
it will be appropriate for the PDs to consult and include the LC.

As PDs are frequently experienced project members with varied and diverse
backgrounds, the Project can benefit most from this collected experience
when PDs are empowered as delegates of the LC.