Raise travel grant limit to $100,000 for 2026
During 2025, we spent over $115,000 on the travel grant program. We had
decided it was OK to go over budget since we spent so little of it in
2024 and we also felt that the 10th-anniversary all-hands was very
important. However, I think it indicates that we need to increase our
limit for 2026.

Importantly, the Foundation will have a separate $50,000 budget
specifically for RustConf.

It is difficult to estimate the expected usage over the next year. For
example, the all-hands may be significantly different than last year.
However, I think this amount represents a good target.